# backend/apidoc/openapi.yaml
openapi: 3.0.0
info:
  description: API Documentation for Meal Planner App
  version: 1.0.0
  title: Meal Planner APIDoc
servers:
  - url: api
...
paths:
...
  /tags/{tagUID}:
    get:
      tags:
        - tags
      summary: Get Tag By UID
      description: Gets a Tag By UID
      operationId: getTagByUID
      parameters:
        - in: path
          name: tagUID
          schema:
            type: string
          required: true
          description: The UID of the Tag
      responses:
        "200":
          description: OK. Tag succesfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      text:
                        type: string
                        example: "Text"
                  message:
                    type: string
                    example: "Tag successfully retrieved"
        "400":
          description: Bad Request. Incomplete Request Body
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Body should contain text!"
        "409":
          description: Conflict. Tag already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Tag already exists"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
    delete:
      tags:
        - tags
      summary: Delete Tag By UID
      description: Delete Tag By UID
      operationId: deleteIngredientByUID
      parameters:
        - in: path
          name: tagUID
          schema:
            type: string
          required: true
          description: The UID of the Tag
      responses:
        "200":
          description: OK. Tag succesfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Tag successfully deleted!"
        "404":
          description: Not Found. Tag does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Tag does not exist"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /tags/{tagText}:
    get:
      tags:
        - tags
      summary: Get Tag By Text
      description: Gets a Tag By Text
      operationId: getTagByText
      parameters:
        - in: path
          name: tagName
          schema:
            type: string
          required: true
          description: The Text of the Tag
      responses:
        "200":
          description: OK. Tag succesfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "http://localhost:8080/api/tags/3PXW8J1QGkQlZgnjOaKv"
                  message:
                    type: string
                    example: "Tag successfully added to database"
        "400":
          description: Bad Request. Incomplete Request Body
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Body should contain text!"
        "409":
          description: Conflict. Tag already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Tag already exists"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /users:
    post:
      tags:
        - users
      summary: Add User
      description: Adds an User
      operationId: addUser
      requestBody:
        description: An User Object
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "username123"
                password:
                  type: string
                  example: "password"
      responses:
        "200":
          description: OK. User succesfully added to database
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "http://localhost:8080/api/users/3PXW8J1QGkQlZgnjOaKv"
                  message:
                    type: string
                    example: "User successfully added to database"
        "400":
          description: Bad Request. Incomplete Request Body
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Body should contain username and password"
        "409":
          description: Conflict. Username already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Username already exists"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
    get:
      tags:
        - users
      summary: Get User By Username and Password
      description: Gets User Info by Username and Password
      operationId: getUserByUsernamePassword
      requestBody:
        description: An User Object
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "username123"
                password:
                  type: string
                  example: "password"
      responses:
        "200":
          description: OK. User succesfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "http://localhost:8080/api/users/DoARzpyNG8mx5Ke6bvnM"
                  message:
                    type: string
                    example: "User successfully retrieved"
        "404":
          description: Not Found. User does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User does not exist"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /users/{userUID}:
    get:
      tags:
        - users
      summary: Get User By UID
      description: Gets User Info by UID
      operationId: getUserByUID
      parameters:
        - in: path
          name: userUID
          schema:
            type: string
          required: true
          description: The UID of the User
      responses:
        "200":
          description: OK. User succesfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      username:
                        type: string
                        example: "username123"
                      youTube:
                        type: string
                        example: "www.youtube.com/username123"
                      instagram:
                        type: string
                        example: "www.instagram.com/username123"
                      facebook:
                        type: string
                        example: "www.facebook.com/username123"
                      twitter:
                        type: string
                        example: "www.twitter.com/username123"
                      tikTok:
                        type: string
                        example: "www.tikTok.com/username123"
                      reddit:
                        type: string
                        example: "www.reddit.com/username123"
                    createdAt:
                      type: string
                      format: date-time
                      example: "2022-03-22T12:34:52.294Z"
                  message:
                    type: string
                    example: "User successfully retrieved"
        "404":
          description: Not Found. User does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User does not exist"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
    put:
      tags:
        - users
      summary: Update User By UID
      description: Update User Info By UID
      operationId: updateUserByUID
      parameters:
        - in: path
          name: userUID
          schema:
            type: string
          required: true
          description: The UID of the User
      requestBody:
        description: New Password and Old Password. Username cannot be changed.
        content:
          application/json:
            schema:
              type: object
              properties:
                oldPassword:
                  type: string
                  example: "password"
                newPassword:
                  type: string
                  example: "password1"
              required:
                - oldPassword
      responses:
        "200":
          description: OK. User succesfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "http://localhost:8080/api/users/3PXW8J1QGkQlZgnjOaKv"
                  message:
                    type: string
                    example: "User successfully updated!"
        "400":
          description: Bad Request. Incomplete or wrong request body
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Body must contain old password for updating user!"
        "404":
          description: Not Found. User does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User does not exist"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
    delete:
      tags:
        - users
      summary: Delete User By UID
      description: Delete User By UID
      operationId: deleteUserByUID
      parameters:
        - in: path
          name: userUID
          schema:
            type: string
          required: true
          description: The UID of the User
      requestBody:
        description: Password for confirmation
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: "password"
      responses:
        "200":
          description: OK. User succesfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User successfully deleted!"
        "400":
          description: Bad Request. Incomplete or wrong request body
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Body must contain password for deleting user!"
        "404":
          description: Not Found. User does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User does not exist"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"